Added New Controllers - Fixed Registration Page
