+++ demo/src/main/resources/templates/register.html
@@ -119,7 +119,7 @@
     <div class="glass-container">
         <h1>Register New User</h1>
         <!-- Registration form -->
-        <form th:action="@{/register}" method="post" th:object="${user}">
+        <form th:action="@{/users/register}" method="post" th:object="${user}">
             <!-- Username input field -->
             <div class="form-group">
                 <label for="username">Username:</label>
